isCreatedOrder replaced, added history
