+++ src/components/admin/layouts/check-admin-layout/check-admin-layout.css
@@ -103,6 +103,10 @@
     font-family: 'Montserrat Medium';
 }
 
+.check-agreement_no-margin {
+    margin-top: 0;
+}
+
 .check-agreement__title {
     font-size: 18px;
     margin-bottom: 5px;
